# fix(wash-cli): remove double event wait for scale