## Feature or Problem
This PR fixes an issue where the `scale` command waited for the event twice, essentially always timing out.

Here's a before and after of how long a scale operation takes to work:
```
# Before (5.2s)
➜ time wash scale component NDIL5CSMAKAFA3H6AKFYKUWNHYI3UFDT622LMAQTPALZAGES3HSC67FH ghcr.io/brooksmtownsend/http-hello-world-rust:0.1.0 foobar -c 27      

Failed to scale component [foobar] (ref: [ghcr.io/brooksmtownsend/http-hello-world-rust:0.1.0]) on host [NDIL5CSMAKAFA3H6AKFYKUWNHYI3UFDT622LMAQTPALZAGES3HSC67FH]

Caused by:
    Timed out waiting for applicable event, operation may have failed
wash scale component NDIL5CSMAKAFA3H6AKFYKUWNHYI3UFDT622LMAQTPALZAGES3HSC67FH  0.09s user 0.02s system 1% cpu 5.224 total

# After (0.5s)
➜ time ../../target/debug/wash scale component NDIL5CSMAKAFA3H6AKFYKUWNHYI3UFDT622LMAQTPALZAGES3HSC67FH ghcr.io/brooksmtownsend/http-hello-world-rust:0.1.0 foobar -c 26

Component [foobar] (ref: [ghcr.io/brooksmtownsend/http-hello-world-rust:0.1.0]) scaled on host [NDIL5CSMAKAFA3H6AKFYKUWNHYI3UFDT622LMAQTPALZAGES3HSC67FH] to 26 max concurrent instances
../../target/debug/wash scale component   foobar -c 26  0.40s user 0.03s system 76% cpu 0.556 total
```

Put another way, this PR makes `wash scale component` 10x faster! 😄 

## Related Issues
<!--- 
Link to any issues or correlated pull requests that are related to this PR. For example, if this PR fixes an issue, link to that issue here.
--->

## Release Information
wash-cli 0.30.1

## Consumer Impact
<!---
Indicate the impact, if any, this change will have on other consumers, dependencies, or dependents. In other words, the "blast radius" of the impact of this change and what steps related projects may need to take in response to this.
--->

## Testing
<!---
Declare the testing information for this pull request
--->

### Unit Test(s)
<!---
Indicate if unit tests were added or modified, and if so, which ones 
--->

### Acceptance or Integration
<!---
Indicate any changes or additions to the acceptance or integration test suite 
--->

### Manual Verification
<!---
Mandatory. Indicate the steps that you took to verify that this pull request works 
--->